QcomModulePkg: Ensure 4KB Alignment for PVM Firmware Size
When the bootloader is unlocked, the PVM firmware size is set to the
partition size. However, when the device is locked, the size is set to
the image size. This causes an RM error if the image size is not aligned
to 4KB.

Fixed this issue by always using the image size with 4KB alignment.

The patch also adds comments and renames some variables for
better readability.

Change-Id: I0927d2ee7c74ad3672b7da408f385ab6cbdd3067